Change processing functions.
